requirements.txt
alpaca-trade-api>=3.2.0
anthropic>=0.18.1
python-dotenv>=1.0.0
pandas>=2.0.0
pytest>=7.4.0
